Add the ability to skip rate limits via a skip_function
This is useful to skip the application of certain limits based on some
aspects of the request e.g. if it is authenticated or from a risky
location.

The skip_function allows the reason for skipping to be defined by the
user of Quart-Rate-Limiter however they'd like.

The name Skip follows the express limiter naming as I think it is the
clearest (defaults to not skip when not present).